Debuffed and pricer update
Updated Debuffed with the -ja spells line and it should now get the debuffs on aoe spells for all the targets it hit, also change its name to Debuffing so it doesn't conflict with the official windower addon.

Updated pricer to be able to do https requests, it requires de addition of new libraries, read the READ THIS.txt for where to place those files.